initial_conditionとboundary_conditionに分けていたものを再結合した。
